update libsodium
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@38182 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/libsodium/src/libsodium/crypto_onetimeauth/poly1305/donna/poly1305_donna32.h
@@ -28,7 +28,7 @@ typedef struct poly1305_state_internal_t {
 static void
 poly1305_init(poly1305_state_internal_t *st, const unsigned char key[32])
 {
-    /* r &= 0xffffffc0ffffffc0ffffffc0fffffff */
+    /* r &= 0xffffffc0ffffffc0ffffffc0fffffff - wiped after finalization */
     st->r[0] = (LOAD32_LE(&key[0])) & 0x3ffffff;
     st->r[1] = (LOAD32_LE(&key[3]) >> 2) & 0x3ffff03;
     st->r[2] = (LOAD32_LE(&key[6]) >> 4) & 0x3ffc0ff;
@@ -225,10 +225,10 @@ poly1305_finish(poly1305_state_internal_t *st, unsigned char mac[16])
     f  = (unsigned long long) h3 + st->pad[3] + (f >> 32);
     h3 = (unsigned long) f;
 
-    STORE32_LE(mac + 0, h0);
-    STORE32_LE(mac + 4, h1);
-    STORE32_LE(mac + 8, h2);
-    STORE32_LE(mac + 12, h3);
+    STORE32_LE(mac + 0, (uint32_t) h0);
+    STORE32_LE(mac + 4, (uint32_t) h1);
+    STORE32_LE(mac + 8, (uint32_t) h2);
+    STORE32_LE(mac + 12, (uint32_t) h3);
 
     /* zero out the state */
     sodium_memzero((void *) st, sizeof *st);
